Fixed issue #5831: User name not properly quoted on display Fixed issue #5832: Unsanitized variable id when viewing a single response

+++ admin/browse.php
@@ -10,14 +10,13 @@
  * other free or open source software licenses.
  * See COPYRIGHT.php for copyright notices and details.
  *
- * $Id: browse.php 12164 2012-01-23 14:24:20Z tmswhite $
 */
 
 include_once("login_check.php");  //Login Check dies also if the script is started directly
 
-if (!isset($limit)) {$limit=returnglobal('limit');}
+if (!isset($limit)) {$limit=(int)returnglobal('limit');}
 if (!isset($surveyid)) {$surveyid=returnglobal('sid');}
-if (!isset($id)) {$id=returnglobal('id');}
+if (!isset($id)) {$id=(int)returnglobal('id');}
 if (!isset($order)) {$order=returnglobal('order');}
 if (!isset($browselang)) {$browselang=returnglobal('browselang');}
